[FIX] stock: We assign partne_id to returned outgoing stock moves from the picking partner_id
'Deliver Slip' report needs to have the delivery address if it was a return of receiving a purchase order

Steps to reproduce:
1. Install Inventory and Purchase Apps
2. Create a PO.
3. Recieve the PO in your warehouse
4. Create a return from the recieve
5. Print the 'Delivery slip'

Solution:
The 'Delivery Slip' report does not know it should print the delivery address because the stock moves doesn't have a 'partner_id'. This is because the values are copied from the stock moves of the incoming picking, So the solution is that while copying the values we add the 'partner_id' value from the picking.

OPW-3623451